+++ src/handlers/index.ts
@@ -1,2 +1,1 @@
-export * as petePlanHandlers from './pete-plan.handlers';
-export * as workoutHandlers from './workout.handlers';
+export * from './workout.handlers.js';